uiディレクトリを作成
